==> 4. Get data redy to be jsonify.ipynb <==
import pandas as pd
import sqlite3
from sqlite3 import Error
conn = sqlite3.connect('OldPortsApp/db/Oldports.sqlite')
c = conn.cursor()


# Load Games
games = "OldPortsApp/db/OldportsGeoTagData.csv"
games = pd.read_csv(games)
games.to_sql('GeoTagData', conn, if_exists = "replace")


#test
df = pd.read_sql('Select * from GeoTagData', conn)
df.columns


import os
 
import pandas as pd
import numpy as np
import sqlite3 as sql
import seaborn as sns
import matplotlib.pyplot as plt

import sqlalchemy 
from sqlalchemy.ext.automap import automap_base
from sqlalchemy.orm import Session
from sqlalchemy import create_engine
from flask import Flask, jsonify, render_template
from flask_sqlalchemy import SQLAlchemy

database_path = "OldPortsApp/db/Oldports.sqlite"
engine = create_engine(f"sqlite:///{database_path}")

Base = automap_base() # Declare a Base using `automap_base()`
Base.prepare(engine, reflect=True) # Creates a class for each table
Base.classes.keys() # display the mapped classses, the tables converted into classes

# print (engine.table_names()) 

data = pd.read_sql("SELECT * FROM GeoTagData", conn) #import pandas

data.head(1)

index=0
df=data.loc[(data["index"] == index) ]
piclocation={'FileAddress':list(df['FileAddress'])[0]}
piclocation["FileAddress"]
